Fixed sdf_write_namevalue when h%string_length changes.
Care must be taken when writing the array of strings in the
sdf_write_namevalue routine since the strings are of length
c_max_string_length which might be longer than the current
value of h%string_length.

It is also necessary to truncate the strings being passed into
this routine in order to prevent warning messages being printed
to screen.